requirements.txt
numpy
scipy
pandas
ipython
jupyter
matplotlib
scikit-image
scikit-learn
sphinx
pytest
pytest-pep8
pytest-cov
requests
pyqt5
flask
gunicorn
timestring
sqlalchemy
Pillow